Add files via upload
How to create a Reserved Word model rule via SQL, along with the rule the instructions created.   Enforce Naming Standard for reserved words does not work on both entity and attributes.   So you either have to run setup twice and change the values in XML, or import the rule, rename it, and reconfigure it to work on Entity names.