Add files via upload
ERWin allows for uploading a file of name translations.   For example, this will translate "Purchase Order" to "PO", according the translate file.

I created this naming standard from an old file I used in ERWin; I uploaded it into a named pair table, and created the XML in the same way I did for reserved words.   I have that file and the XML in a SQL statement.  Let me know if you'd like it and I'll dig it up and post it here.